Add support for HTTP Basic authentication
For normal setups this is not required. However, when using external
authentication (e.g. in setups using LDAP authentication), the API
endpoint might only be accessible after the client authenticated. For
that we add the new parameters `auth_username` and `auth_password`. If
both of these are set the API calls will be made using these
credentials.