Configure SSL certificate (#26)
* Test update for ssl_verify in config

* Configure SSL certificate

The user now can set whether to use an SSL certificate or not, or even
set their own self-signed certificate from the config file. The config
file comes with additional instruction and warning about SSL
certificates.

If ssl_verify isn't set in the config file, it defaults to true.

* Code autoformat

* Fix config file permission

User config file contains a secret API that shouldn't be readable to
others

* typo

* Moved file permission change to appropriate function